Add GitHub Action Configuration (#18)
# Pull Request

## Description

This pull request introduces a new GitHub Action named "Repository
Visualiser". The most important changes include adding metadata for the
action and specifying the runtime environment.

Metadata for the action:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R1-R10):
Added the name, description, and author information for the "Repository
Visualiser" action.

Runtime environment:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R1-R10):
Configured the action to use Docker, specifying the `Dockerfile` and
setting the `GITHUB_ACTION` environment variable to "true".

Fixes #14