Ignore Files and Folders from Visualisation (#26)
# Pull Request

## Description

This pull request introduces functionality to filter out files and
directories based on an ignore list defined in a JSON file. The changes
include adding the necessary JSON parsing, updating the main function to
apply the filtering, and implementing the filtering logic.

### Enhancements to file filtering:

*
[`ignore.json`](diffhunk://#diff-87361699ebe16f272a1398d6c1cea58e61bcb72410bd775a238ecf731a195020R1-R6):
Added a new JSON file to specify files and directories to ignore.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R4):
Imported the `encoding/json` package to handle JSON parsing.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R20-R29):
Updated the `main` function to load the ignore list and filter the file
statistics accordingly.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R82-R124):
Added the `loadIgnoreList` function to read and parse the ignore list
from `ignore.json`.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R82-R124):
Implemented the `filterIgnoredFiles` function to exclude files and
directories based on the patterns in the ignore list.

Fixes #25