Add GitHub Action Summary Language Table (#31)
# Pull Request

## Description

This pull request includes several changes to the `main.go` file,
focusing on improving the handling and reporting of language counts in
the GitHub Action summary. The most important changes include adding a
new import, reordering the function calls in `main`, and enhancing the
`writeSummary` function to better handle unknown languages and format
the output.

Enhancements to language count handling and summary reporting:

*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R9):
Added `sort` to the import statements to enable sorting of language
counts.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261L18-R19):
Reordered the `writeSummary` function call in `main` to ensure it is
executed after potential errors are handled.
[[1]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261L18-R19)
[[2]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R34-R83)
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R34-R83):
Enhanced the `writeSummary` function to separate and count unknown
languages, sort known languages by descending file count, and format the
summary output as a markdown table.
*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R297-R299):
Modified the `recursiveFileSearch` function to label files with unknown
languages explicitly as "Unknown".

Fixes #21