Push Diagram to Branch (#37)
# Pull Request

## Description

This pull request includes significant changes to the repository
visualizer GitHub Action, Dockerfile, and main application logic to
enhance functionality and integration with GitHub.

### Enhancements to GitHub Action:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R104-R106):
Added permissions to write contents and pull requests, and included
`GITHUB_TOKEN` and `GITHUB_REPOSITORY` as inputs for the repository
visualizer job.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R104-R106)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R115-R117)
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R5-R19):
Introduced required inputs `GITHUB_TOKEN` and `GITHUB_REPOSITORY` for
the action and updated the environment variables to use these inputs.

### Dockerfile improvements:

*
[`Dockerfile`](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L3-R14):
Switched the base image to `golang:1.23.5-bookworm`, added `ignore.json`
to the copy command, and updated the build command to output to
`/application`.

### Main application changes:

*
[`main.go`](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R4-R25):
Refactored the main function to push the SVG diagram to a branch and
comment on the pull request with the link to the diagram. Added
functions `pushSVGToBranch` and `commentOnPR` to handle these tasks.
[[1]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R4-R25)
[[2]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261L29-R55)
[[3]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R338-R462)

### Dependency updates:

*
[`go.mod`](diffhunk://#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6L5-R15):
Added new dependencies `github.com/google/go-github/v48` and
`golang.org/x/oauth2` for GitHub API interactions.

Fixes #push