Regularly Clear Caches (#55)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to clean
caches on a scheduled basis. The workflow is designed to run monthly and
can also be triggered manually.

New GitHub Actions workflow:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R24):
Added a workflow named "Clean Caches" that schedules a job to run on the
first day of each month and allows manual triggering. The job checks out
the repository and deletes all caches using the GitHub CLI.