Refactor Pull Request Comment (#77)
# Pull Request

## Description

This pull request includes several changes to GitHub Actions
configurations and the `visualiser` tool. The most important changes
involve adding new inputs and environment variables to the GitHub
Actions workflow and simplifying the process of posting SVG images as
comments on pull requests.

### GitHub Actions Configuration:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L64-R65):
Added `action.yml` to the list of files that trigger GitHub Actions
workflows.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R134):
Added `GITHUB_REPOSITORY_OWNER` to the list of environment variables
used in the workflow.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R9-R11):
Added `GITHUB_REPOSITORY_OWNER` as a required input for the action.

### `visualiser` Tool Updates:

*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L35-R36):
Removed the `pushSVGToBranch` function and updated the `commentOnPR`
function to post the SVG image directly as a base64-encoded string in
the pull request comment.
[[1]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L35-R36)
[[2]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L339-R335)
[[3]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L440-L448)
[[4]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96R371-R375)