Update GitHub Action Summary (#83)
# Pull Request

## Description

This pull request includes several changes to the `visualiser`
application and the GitHub Actions workflow to enhance functionality and
improve error handling. The most important changes include adding a new
GitHub Action to upload a generated diagram, improving error handling in
the `main` function, and ensuring valid values in the SVG generation
process.

### Improvements to GitHub Actions:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R152-R156):
Added a new step to upload the generated `repository_visualisation.svg`
diagram as an artifact.

### Enhancements to `visualiser` application:

*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96R26-R29):
Improved error handling in the `main` function by checking if
`fileStats` is empty and printing a message if no files are found.
*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L41-L102):
Modified the `writeSummary` function to include `markdownContent` and
moved the function to a more appropriate location in the file.
[[1]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L41-L102)
[[2]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96R341-R391)
*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L203-R157):
Ensured valid radius values in the `generateSVG` function to avoid zero
or negative values and division by zero.
[[1]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L203-R157)
[[2]](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L245-R206)
*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L41-L102):
Changed the file path for writing the SVG output to
`/github/workspace/repository_visualisation.svg`.

These changes improve the robustness and functionality of the
`visualiser` application and its integration with GitHub Actions.

Fixes #11