Migrate Ignore to GO File (#85)
# Pull Request

## Description

This pull request includes changes to simplify the handling of ignored
files in the `visualiser` package by replacing a dynamic ignore list
loaded from a JSON file with a static default ignore list defined in the
code. The most important changes include the addition of a default
ignore list, the removal of the JSON-based ignore list loading, and the
update to use the default ignore list in the main processing function.

Changes to ignore list handling:

*
[`visualiser/ignore.go`](diffhunk://#diff-86dc458e369a0cad377b45954e0baa89a69be194d8002fdc95cb24719d0f7d71R1-R27):
Added a `defaultIgnoreList` variable containing patterns for files and
folders to ignore.
*
[`visualiser/ignore.json`](diffhunk://#diff-fd76827a2c36936f7941e4d7a37fbe3b0201b68b6a3ce9c63b98a32f4c759cacL1-L3):
Removed the JSON file that previously contained ignore patterns.
*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L31-R31):
Updated the `main` function to use the `defaultIgnoreList` instead of
loading the ignore list from a JSON file.
*
[`visualiser/main.go`](diffhunk://#diff-7ccd8b79b3553c7bd6e5bbcabffa8a3c5192001f95f13f4fbf6a66d17e44ff96L88-L102):
Removed the `loadIgnoreList` function, which was responsible for loading
the ignore list from the JSON file.

Fixes #84