Remove CodeQL Configuration (#96)
# Pull Request

## Description

This pull request updates the CodeQL workflow configuration to use the
latest versions of the CodeQL actions and removes a custom query filter
configuration. The changes aim to streamline the workflow and ensure
compatibility with the latest CodeQL features.

### CodeQL Workflow Updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L109-R109):
Updated the versions of the `upload-sarif`, `init`, and `analyze` CodeQL
actions to `v3.28.15` to ensure the workflow uses the latest features
and fixes.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L109-R109)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L130-R135)

### Configuration Simplification:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Removed the custom query filter for `actions/unpinned-tag`, as it is no
longer required.