Bump astral-sh/setup-uv from to 6.0.0 (#102)
# Pull Request

## Description

This pull request updates the GitHub Actions workflow to use the latest
version of the `uv` setup action. The change ensures compatibility with
the newest features and fixes provided in version 6.0.1 of the
`astral-sh/setup-uv` action.

Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L83-R83):
Updated the `astral-sh/setup-uv` action from version `v5.4.2` to
`v6.0.1` in the `Install the latest version of uv` step.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L83-R83)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L101-R101)